Updated __main__()
Added `if` statements to check if `shared.model_name` is not `None` before using it to access `shared.settings`.
Updated the way default values for `default_preset` and `default_text` are set based on the value of `shared.model_name`.
Simplified the logic for setting `shared.model_name` based on command-line arguments and available models.
Updating the way default extensions are set based on whether the program is running in chat mode or not.